fix verbose, gap space and add email.yaml
