+++ xlswriter/emailaddress.py
@@ -1,69 +1,58 @@
-from emailaddress_incoming_urls import EmailAddress_Incomming_Urls
-from emailaddress_outgoing_urls import EmailAddress_Outgoing_Urls
 from py2neo import Graph
+from logger import Logger
+from processor import Worker
+from xlsxhandler import XlsHanler
 import xlsxwriter
 import os
 import logging
-from logger import Logger
+import yaml
+import sys
+
 
 class EmailAddress():
 
-    def __init__(self, email_address, neo4j_conn, output_path ,sections):
-        self.incomming_urls = EmailAddress_Incomming_Urls(neo4j_conn, email_address)
-        self.outgoing_urls = EmailAddress_Outgoing_Urls(neo4j_conn, email_address)
+    def __init__(self, email_address, neo4j_conn, output_path, conf):
+        self.worker = Worker(neo4j_conn, email_address=email_address)
+        self.xls_handler = XlsHanler(os.path.join(
+            output_path, 'EMAIL_{0}.xlsx'.format(email_address)))
 
-        self.sections = sections
         self.email_address = email_address
         self.output_path = output_path
-        self.data = self.__processor()
+        for tab_config in conf:
+            self.tab_handler(tab_config)
 
-    def __processor(self):
-        data = {}
-        for section in self.sections:
-            if section == 'UrlsIn':
-                data[section] = []
-                for v in self.incomming_urls.get_result():
-                    data[section].append(v)
-            elif section == 'UrlsOut':
-                data[section] = []
-                for v in self.outgoing_urls.get_result():
-                    data[section].append(v)
+    def tab_handler(self, tab_config):
+        row = 2
+        col = 2
+        gap = tab_config['gap']
+        if tab_config['orientation'] == 'horizontal':
+            hoz = True
+        else:
+            hoz = False
 
-        return data
+        if tab_config['autofit'] == 'True':
+            af = True
+        else:
+            af = False
 
-    def write_to_xls(self):
-        output_file = os.path.join(self.output_path, 'EMAIL_{0}.xlsx'.format(self.email_address))
-        workbook = xlsxwriter.Workbook(output_file)
-        for section in self.data.keys():
-            row = 0
-            col = 0
-            worksheet = workbook.add_worksheet(section)
-            if self.data[section]:
-                keys = self.data[section][0].keys()
-                for k in keys:
-                    worksheet.write(row, col, k)
-                    col += 1
-                col = 0
-                row = 1
-                for v in self.data[section]:
-                    for k in keys:
-                        worksheet.write_string(row, col, v[k])
-                        col += 1
-                    col = 0
-                    row += 1
-
-        workbook.close()
-
-    def in_json_format(self):
-        pass
-
-    def in_yaml_format(self):
-        pass
+        for query in tab_config['queries']:
+            data = self.worker.get_result(query['query'])
+            if not data:
+                continue
+            row, col = self.xls_handler.save_data_to_tab(
+                tab_config['name'], data, row, col, hoz, af)
+            if hoz:
+                col += gap
+                row = 2
+            else:
+                row += gap
+                col = 2
 
 
 class EmailAddress_Handler(Logger):
 
-    def __init__(self, output_path, sections,
+    def __init__(self, output_path,
+                 config_file='perspectives/EMAIL.yaml',
                  neo4j_connection_string="bolt://localhost:7687",
                  neo4j_user="neo4j",
                  neo4j_password="neo4j"):
@@ -71,8 +60,14 @@ class EmailAddress_Handler(Logger):
         self.graph = Graph(neo4j_connection_string,
                            user=neo4j_user,
                            password=neo4j_password)
-        self.sections = sections
         self.output_path = os.path.abspath(output_path)
+        if config_file:
+            with open(config_file, 'r') as f:
+                try:
+                    self.conf = yaml.load(f)
+                except Exception as e:
+                    raise e
+                    sys.exit(1)
 
     @property
     def all_emails(self):
@@ -81,9 +76,9 @@ class EmailAddress_Handler(Logger):
 
     def process(self, verbose):
         if verbose:
-            for i in ['EmailAddress_Incomming_Urls', 'EmailAddress_Outgoing_Urls']:
+            for i in ['Worker']:
                 logging.getLogger(i).setLevel(logging.INFO)
 
         for email_address in self.all_emails:
-            d = EmailAddress(email_address, self.graph, self.output_path,self.sections)
-            d.write_to_xls()
+            d = EmailAddress(email_address, self.graph,
+                             self.output_path, self.conf['tabs'])
