+++ xlswriter/perspectives/EMAIL.yaml
@@ -1,0 +1,27 @@
+doc:
+    type: EMAIL
+    format: xls
+
+tabs:
+    - name: Incomming Url
+      header: plain
+      gap: 2
+      orientation: vertial
+      autofit: true
+      queries:
+          - title: Urls
+            query: MATCH (m:Message)-[:TO]->(e:EmailAddress), (m)-[:CONTAINS]->(u:Url) WHERE e.name = "{email_address}" RETURN u.full_link as url
+            show_labels: true
+
+    - name: Outgoing Url
+      header: full
+      gap: 2
+      orientation: vertial
+      autofit: true
+      queries:
+          - title: Urls
+            query: MATCH (e:EmailAddress)-[:SENT]->(m:Message), (m)-[:CONTAINS]->(u:Url) WHERE e.name = "{email_address}" RETURN u.full_link as url
+            show_labels: false
+
+
+
