+++ xlswriter/run_from_disk.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
 PATH_DATA="~/tmp/allmail"
-python main.py --perspective DOMAIN --sections 'Email Address,Urls' --transport FILE --file_store_path $PATH_DATA
+python main.py --perspective DOMAIN --transport FILE --file_store_path $PATH_DATA
